requirements.txt
aiofiles==0.5.0
alembic==1.6.5
aniso8601==7.0.0
arrow==1.2.1
async-exit-stack==1.0.1
async-generator==1.10
attrs==21.2.0
autopep8==1.5.7
bcrypt==3.2.0
binaryornot==0.4.4
certifi==2021.5.30
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.4
click==7.1.2
colorama==0.4.4
cookiecutter==1.7.3
cryptography==3.4.8
dnspython==2.1.0
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.68.0
graphene==2.1.9
graphql-core==2.3.2
graphql-relay==2.0.1
greenlet==1.1.1
h11==0.12.0
httptools==0.1.2
idna==3.2
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
jinja2-time==0.2.0
Mako==1.1.5
MarkupSafe==2.0.1
orjson==3.6.4
packaging==21.3
passlib==1.7.4
pluggy==1.0.0
poyo==0.5.0
promise==2.3
psycopg2==2.9.2
py==1.11.0
pyasn1==0.4.8
pycodestyle==2.7.0
pycparser==2.20
pydantic==1.8.2
pyparsing==3.0.6
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==0.19.0
python-editor==1.0.4
python-jose==3.3.0
python-multipart==0.0.5
python-slugify==5.0.2
PyYAML==5.4.1
requests==2.26.0
rsa==4.7.2
Rx==1.6.1
six==1.16.0
SQLAlchemy==1.4.23
starlette==0.14.2
text-unidecode==1.3
toml==0.10.2
typing-extensions==3.10.0.0
ujson==4.0.2
urllib3==1.26.6
uvicorn==0.13.4
uvloop==0.16.0
watchgod==0.7
websockets==10.1
